Treat get_lock(), is_free_lock(), and is_used_lock() as TYPE_WRITE
If an operation requires a lock, then it should be on the read-write master
-- because the lock-system is for use with all concurrent processes.

Relatedly, is_free_lock() and is_used_lock() determine if the *anyone* holds
a lock on a resource. That needs to be checked centrally.

What about release_lock() and release_all_locks()?  These only release locks
held by the current MySQL session.  Two possibilities: (1) we've had a
preceding call to get_lock() (in which case we've already switched to
read-write master), or (2) we've never called get_lock (in which case
there's nothing releaseable).  Either way, we never need to force a
reconnect on account of release_lock().  No point wasting compute cycles on
scanning for it.